Created the following:
New reducer function in the state.
New useEffect hook on a component to query my DB.
My DB is empty, so unable to check if the useEffect fetch funciton is working. Started with creating a recipe to POST to DB, but ran into an issue with a component. Finally debugged it, will resume tomorrow.